Split spacesView and spacesStorageView
