+++ components/experimental/spaces/src/spacesStorageView.css
@@ -1,4 +1,4 @@
-.spaces-component-grid {
+.spaces-storage-view {
     padding-top: 7rem;
     min-height: 3000px;
     width: 100%;
